Probando

+++ Investments/src/com/miguel/HYIP/Motor/Avisador.java
@@ -6,7 +6,7 @@ public class Avisador {
 
 	public static void main(String[] args) {
 		HourlyBank hourlyBank = new HourlyBank();
-		long amount = 0;
+		double amount = 0;
 		
 		if (hourlyBank.isAlive()) {
 			hourlyBank.login();
